Move the helper function to their own module
